+++ app/src/main/java/com/pixelnos/fire/backend/manager/CyclicalTransaction.java
@@ -1,4 +1,4 @@
-package com.pixelnos.fire.backend.expenseManager;
+package com.pixelnos.fire.backend.manager;
 
 import java.util.Calendar;
 import java.util.Date;
