+++ app/src/main/java/com/pixelnos/fire/backend/manager/Transaction.java
@@ -1,4 +1,4 @@
-package com.pixelnos.fire.backend.expenseManager;
+package com.pixelnos.fire.backend.manager;
 import java.util.ArrayList;
 import java.util.Date;
