+++ app/src/test/java/com/pixelnos/fire/backend/manager/CurrencyTest.java
@@ -1,4 +1,4 @@
-package com.pixelnos.fire.backend.expenseManager;
+package com.pixelnos.fire.backend.manager;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
