+++ app/src/test/java/com/pixelnos/fire/backend/manager/PeriodTest.java
@@ -1,4 +1,4 @@
-package com.pixelnos.fire.backend.expenseManager;
+package com.pixelnos.fire.backend.manager;
 import static org.junit.Assert.assertEquals;
 import org.junit.Test;
